requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
discord==2.3.2
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
numpy==2.2.6
pillow==11.3.0
propcache==0.3.2
pyboy==2.6.0
PySDL2==0.9.17
typing_extensions==4.14.1
yarl==1.20.1
